Move TestConfig to `tests/testing_utils.py` and remove `tests/data.py`
